requirements.txt
duckduckgo_search
google-generativeai
